+++ guillotina/documentation/sphinx/__init__.py
@@ -88,26 +88,35 @@ class HTTPService(CodeBlock):
         service = data['service']
         name = service.get('name') or ''
         path_scheme = data.get('path_scheme') or name
-        description = service.get('description') or ''
+        summary = service.get('summary') or ''
         permission = service.get('permission') or ''
 
         container = nodes.container('')
-        container.append(nodes.header())
-        container.append(addnodes.desc_name(method + ' ', method + ' '))
-        container.append(addnodes.desc_name(path_scheme, path_scheme))
+        container.append(addnodes.desc_name('', method + ' '))
+        container.append(addnodes.desc_name('', path_scheme))
 
         inner_container = nodes.definition('')
         container.append(inner_container)
 
-        inner_container.append(nodes.paragraph(description, description))
+        inner_container.append(nodes.paragraph(summary, summary))
         inner_container.append(addnodes.desc_name('permission', 'permission'))
         perm_label = ': ' + permission
         inner_container.append(addnodes.desc_annotation(perm_label, perm_label))
         inner_container.append(example.run()[0])
 
+        # extra = nodes.paragraph('', '')
+        # inner_container.append(extra)
+        # if service.get('responses'):
+        #     extra.append(nodes.strong('', 'Responses'))
+        #     blist = nodes.bullet_list('')
+        #     extra.append(blist)
+        #     for code, config in service['responses'].items():
+        #         blist.append(render_response(code, 'Hello'))
+
         # cleanup
         os.remove(request_filename)
         os.remove(response_filename)
+
         return container
 
     def run(self):
@@ -117,7 +126,12 @@ class HTTPService(CodeBlock):
             if filename.startswith(type_name + '-'):
                 files.append(filename)
         files.sort(key=service_filename_sort_key)
-        return [self.process_service(filename) for filename in files]
+
+        env = self.state.document.settings.env
+        targetid = "service-%d" % env.new_serialno('service')
+        targetnode = nodes.target('', '', ids=[targetid])
+
+        return [targetnode] + [self.process_service(filename) for filename in files]
 
 
 def setup(app):